# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProofBase API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product listing operations
  - name: feedback
    description: Feedback operations
  - name: dashboard
    description: Founder dashboard operations
  - name: ai
    description: AI-powered insights
  - name: leaderboard
    description: Leaderboard and rankings
  - name: feed
    description: Community activity feed
  - name: gamification
    description: Points, badges, and streaks
  - name: geo
    description: Geocoding utilities
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, newest, top_rated, score]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/trending:
    get:
      operationId: getTrendingProducts
      tags: [products]
      summary: Get trending products
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/map:
    get:
      operationId: getMapProducts
      tags: [products]
      summary: Get all products with location data for map display
      responses:
        "200":
          description: Products with location
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapProduct"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  /products/{id}/feedback:
    get:
      operationId: getProductFeedback
      tags: [feedback]
      summary: Get all feedback for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackBody"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  /products/{id}/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

  /products/{id}/summarize:
    post:
      operationId: summarizeProductFeedback
      tags: [ai]
      summary: Generate AI summary of product feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI-generated feedback summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSummary"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get founder dashboard data
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"

  /dashboard/products:
    get:
      operationId: getDashboardProducts
      tags: [dashboard]
      summary: Get all products owned by the authenticated founder
      responses:
        "200":
          description: Founder products with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWithStats"

  /leaderboard/builders:
    get:
      operationId: getTopBuilders
      tags: [leaderboard]
      summary: Get top founders ranked by product quality and engagement
      responses:
        "200":
          description: Top builders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardBuilder"

  /leaderboard/reviewers:
    get:
      operationId: getTopReviewers
      tags: [leaderboard]
      summary: Get top reviewers ranked by feedback count
      responses:
        "200":
          description: Top reviewers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardReviewer"

  /leaderboard/products:
    get:
      operationId: getTopProducts
      tags: [leaderboard]
      summary: Get top products by weighted score
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /feed:
    get:
      operationId: getCommunityFeed
      tags: [feed]
      summary: Get community activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Community feed events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedEventItem"

  /gamification/me:
    get:
      operationId: getMyPoints
      tags: [gamification]
      summary: Get current user's points, streak, and badges
      responses:
        "200":
          description: User gamification data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGamification"

  /gamification/me/feedback-count:
    get:
      operationId: getMyFeedbackCount
      tags: [gamification]
      summary: Get number of feedbacks given by current user
      responses:
        "200":
          description: Feedback count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackCount"

  /gamification/checkin:
    post:
      operationId: dailyCheckin
      tags: [gamification]
      summary: Daily login check-in for streak and points
      responses:
        "200":
          description: Updated gamification data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGamification"

  /geo/search:
    get:
      operationId: geocodeSearch
      tags: [geo]
      summary: Search for a location and return coordinates
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Geocoding results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeoResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        founderClerkId:
          type: string
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        category:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        feedbackCount:
          type: integer
        avgRating:
          type: ["number", "null"]
        score:
          type: ["number", "null"]
        statusTag:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - founderClerkId
        - name
        - tagline
        - description
        - category
        - feedbackCount
        - createdAt
        - updatedAt

    MapProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tagline:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
      required:
        - id
        - name
        - tagline
        - latitude
        - longitude

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        category:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
      required:
        - name
        - tagline
        - description
        - category

    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        websiteUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        category:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]

    Feedback:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerClerkId:
          type: ["string", "null"]
        liked:
          type: string
        confused:
          type: string
        missing:
          type: string
        wouldPay:
          type: boolean
        rating:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - productId
        - liked
        - confused
        - missing
        - wouldPay
        - rating
        - createdAt

    CreateFeedbackBody:
      type: object
      properties:
        liked:
          type: string
        confused:
          type: string
        missing:
          type: string
        wouldPay:
          type: boolean
        rating:
          type: integer
      required:
        - liked
        - confused
        - missing
        - wouldPay
        - rating

    ProductStats:
      type: object
      properties:
        productId:
          type: integer
        totalFeedback:
          type: integer
        avgRating:
          type: ["number", "null"]
        wouldPayCount:
          type: integer
        wouldPayRatio:
          type: ["number", "null"]
        ratingDistribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingBucket"
      required:
        - productId
        - totalFeedback
        - wouldPayCount

    RatingBucket:
      type: object
      properties:
        rating:
          type: integer
        count:
          type: integer
      required:
        - rating
        - count

    FeedbackSummary:
      type: object
      properties:
        productId:
          type: integer
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        overallSentiment:
          type: string
        generatedAt:
          type: string
      required:
        - productId
        - strengths
        - weaknesses
        - suggestions
        - overallSentiment
        - generatedAt

    DashboardData:
      type: object
      properties:
        totalProducts:
          type: integer
        totalFeedback:
          type: integer
        avgRating:
          type: ["number", "null"]
        recentFeedback:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackWithProduct"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductWithStats"
        feedbackOverTime:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackTimePoint"
      required:
        - totalProducts
        - totalFeedback
        - recentFeedback
        - topProducts

    FeedbackTimePoint:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count

    FeedbackWithProduct:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        liked:
          type: string
        confused:
          type: string
        missing:
          type: string
        wouldPay:
          type: boolean
        rating:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - liked
        - confused
        - missing
        - wouldPay
        - rating
        - createdAt

    ProductWithStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tagline:
          type: string
        category:
          type: string
        logoUrl:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        feedbackCount:
          type: integer
        avgRating:
          type: ["number", "null"]
        wouldPayRatio:
          type: ["number", "null"]
        score:
          type: ["number", "null"]
        statusTag:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - tagline
        - category
        - feedbackCount
        - createdAt

    LeaderboardBuilder:
      type: object
      properties:
        userId:
          type: string
        productCount:
          type: integer
        totalFeedback:
          type: integer
        avgRating:
          type: ["number", "null"]
        totalPoints:
          type: integer
        badges:
          type: array
          items:
            type: string
        topProduct:
          type: ["string", "null"]
      required:
        - userId
        - productCount
        - totalFeedback
        - totalPoints
        - badges

    LeaderboardReviewer:
      type: object
      properties:
        userId:
          type: string
        feedbackCount:
          type: integer
        totalPoints:
          type: integer
        badges:
          type: array
          items:
            type: string
      required:
        - userId
        - feedbackCount
        - totalPoints
        - badges

    FeedEventItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - type
        - createdAt

    UserGamification:
      type: object
      properties:
        userId:
          type: string
        points:
          type: integer
        streak:
          type: integer
        lastActiveDate:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string
      required:
        - userId
        - points
        - streak
        - badges

    FeedbackCount:
      type: object
      properties:
        count:
          type: integer
        canListProduct:
          type: boolean
      required:
        - count
        - canListProduct

    GeoResult:
      type: object
      properties:
        displayName:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - displayName
        - country
        - latitude
        - longitude